+++ nctest/add.c
@@ -56,7 +56,7 @@ add_var (struct netcdf *test, struct cdfvar *ivar) /* add the variable ivar to t
     (void) strcpy(test->vars[test->nvars].name, ivar->name);
     test->vars[test->nvars].type = ivar->type;
     test->vars[test->nvars].ndims = ivar->ndims;
-    test->vars[test->nvars].dims = (int *) emalloc(sizeof(int)*ivar->ndims);
+    test->vars[test->nvars].dims = (int *) emalloc(sizeof(int)*(size_t)ivar->ndims);
     for (i = 0; i < ivar->ndims; i++)
       test->vars[test->nvars].dims[i] = ivar->dims[i];
     test->vars[test->nvars].natts = 0;
