+++ nctest/varputg.c
@@ -65,7 +65,7 @@ test_ncvarputg(const char *path) /* name of writable netcdf file to open */
 	}
 
 	/* fill in vals with value of appropriate type */
-	hc.vals = emalloc(nctypelen(test.vars[iv].type));
+	hc.vals = emalloc((size_t)nctypelen(test.vars[iv].type));
 	val_fill(test.vars[iv].type, 1, hc.vals);
 
 #	define TEST_FAILS(varid) \
