+++ nczarr_test/ut_util.c
@@ -4,6 +4,7 @@
  */
 
 #include "ut_includes.h"
+#include <stddef.h>
 
 #undef DEBUG
 
@@ -81,8 +82,8 @@ parsedimdef(const char* s0, Dimdef** defp)
     s = s0;
     if((p = strchr(s,'=')) == NULL) abort();
     if((count = (p - s)) == 0) return THROW(NC_EINVAL);
-    def->name = malloc(count+1);
-    memcpy(def->name,s,count);
+    def->name = malloc((size_t)count+1);
+    memcpy(def->name,s,(size_t)count);
     def->name[count] = '\0';
     s = p+1;
     sscanf(s,"%u%n",&l,&nchars);
@@ -115,7 +116,7 @@ parsevardef(const char* s0, NClist* dimdefs, Vardef** varp)
     if(p == NULL) return THROW(NC_EINVAL);
     len = (p - s);  
     if(len == 0) return THROW(NC_EINVAL);
-    memcpy(name,s,len);
+    memcpy(name,s,(size_t)len);
     name[len] = '\0';
     vd->typeid = ut_typeforname(name);
     vd->typesize = ut_typesize(vd->typeid);
@@ -127,7 +128,7 @@ parsevardef(const char* s0, NClist* dimdefs, Vardef** varp)
     if(p == NULL) return THROW(NC_EINVAL);
     len = (p - s);  
     if(len == 0) return THROW(NC_EINVAL);
-    memcpy(name,s,len);
+    memcpy(name,s,(size_t)len);
     name[len] = '\0';
     vd->name = strdup(name);     
     /* parse a vector of dimnames and chunksizes and convert */
@@ -178,7 +179,7 @@ parsestringvector(const char* s0, int stopchar, char*** namesp)
     /* First, compute number of elements */
     for(s=s0,nelems=1;*s;s++) {if(*s == ',') nelems++; if(*s == stopchar) break;}
     if(nelems == 0) return THROW(NC_EINVAL);
-    names = calloc(nelems+1,sizeof(char*));
+    names = calloc((size_t)nelems+1,sizeof(char*));
     for(s=s0,i=0;i<nelems;i++) {
         ptrdiff_t len;
         const char* p = strchr(s,',');
@@ -187,8 +188,8 @@ parsestringvector(const char* s0, int stopchar, char*** namesp)
         if(names[i] == NULL) {
             char* q;
             len = (p - s);
-            q = malloc(1+len);
-            memcpy(q,s,len);
+            q = malloc(1+(size_t)len);
+            memcpy(q,s,(size_t)len);
             q[len] = '\0';
             names[i] = q;
         }
@@ -211,7 +212,7 @@ parseintvector(const char* s0, int typelen, void** vectorp)
         if(*s == ',') nelems++;
     }
 
-    vector = calloc(nelems,typelen);
+    vector = calloc((size_t)nelems, (size_t)typelen);
 
     /* Extract the elements of the vector */
     /* Skip any leading bracketchar */
@@ -508,8 +509,8 @@ fillcommon(struct Common* common, Vardef* var)
     common->typesize = sizeof(int);
     if(var != NULL) {
         common->rank = var->rank;
-        memcpy(common->dimlens,var->dimsizes,sizeof(size64_t)*common->rank);
-        memcpy(common->chunklens,var->chunksizes,sizeof(size64_t)*common->rank);
-        memcpy(common->memshape,common->dimlens,sizeof(size64_t)*common->rank); /* fake it */
+        memcpy(common->dimlens,var->dimsizes,sizeof(size64_t)*(size_t)common->rank);
+        memcpy(common->chunklens,var->chunksizes,sizeof(size64_t)*(size_t)common->rank);
+        memcpy(common->memshape,common->dimlens,sizeof(size64_t)*(size_t)common->rank); /* fake it */
     }
 }
